requirements.txt
pandas
numpy
scikit-learn
gensim
nltk
matplotlib
seaborn
plotly
kaleido
wordcloud
textblob
vaderSentiment
scipy
umap-learn
jupyter
tqdm
